Add updated date to podcasts and interviews
I think it's important to show both the published and updated dates. I
also decided to simplify the date formatting. Now it just shows the day
as a number such as 19 instead of 19th.

I also reduced the month name to its 3 letter abbreviation instead of
the full name. With both a date and updated date the line got too
cluttered and wrapped on mobile devices.