Replace "let" with "var" for all JavaScript files
This was causing mayhem when running:
  JEKYLL_ENV=production bundle exec jekyll build

I guess Jekyll or maybe the Jekyll-Assets plugin doesn't know how to
deal with that keyword. Without this change Jekyll would throw an error
such as:

Liquid Exception: SyntaxError: Unexpected token: name
(isProgressSliderDrag) in /_layouts/base.html

But it would only throw this with JEKYLL_ENV set to production.