Add support for optional transcripts
Moving forward I will be using transcripts instead of the 1 line summary
show notes that I was doing prior to this commit.

At the moment I'm paying for a service to have them machine generated
but there's a chance having poor grammar and typos will negatively
affect page rankings so I developed a way to only show human reviewed
transcripts and hopefully make it easy for folks to improve them.