First iteration of supporting multiple guests
This isn't meant to be a final solution. At the moment only episode 46
has multiple guests and I wanted a quick and dirty solution that lets 2
guests be displayed in multiple areas of the site.

I don't think I'm going to have 3+ guests any time soon.

A better solution would be to make guests an array and then each guest
would have their own twitter, and the same code path in item.html could
operate on a list of 1 or more quests, with proper CSS columns.

But that would have involved going back and editing every other podcast
epsiode recorded in the past. If more multi-guests episode happen in the
future maybe I'll do this in the long run.