Fix mp3_bytes and mp3_duration values
These are now auto-calculated by a script that uses ffmpeg to get the
duration and wc -c to get the size of the file in bytes.

Previous to this I was using the Windows properties of the file and
occasionally used the size on disk instead of the size by accident, as
well as having the properties round the duration inconsistently.